sphinx-agent: Reorganise some types
