+++ lib/esbonio/esbonio/sphinx_agent/types/roles.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import re
 import typing
 from dataclasses import dataclass
 from dataclasses import field
@@ -13,6 +14,119 @@ if typing.TYPE_CHECKING:
     from typing import Tuple
 
     from .lsp import Location
+
+MYST_ROLE: re.Pattern = re.compile(
+    r"""
+    ([^\w`]|^\s*)                     # roles cannot be preceeded by letter chars
+    (?P<role>
+      {                               # roles start with a '{'
+      (?P<name>[:\w-]+)?              # roles have a name
+      }?                              # roles end with a '}'
+    )
+    (?P<target>
+      `                               # targets begin with a '`' character
+      ((?P<alias>[^<`>]*?)<)?         # targets may specify an alias
+      (?P<modifier>[!~])?             # targets may have a modifier
+      (?P<label>[^<`>]*)?             # targets contain a label
+      >?                              # labels end with a '>' when there's an alias
+      `?                              # targets end with a '`' character
+    )?
+    """,
+    re.VERBOSE,
+)
+"""A regular expression to detect and parse parial and complete roles.
+
+I'm not sure if there are offical names for the components of a role, but the
+language server breaks a role down into a number of parts::
+
+                 vvvvvv label
+                v modifier(optional)
+               vvvvvvvv target
+   {c:function}`!malloc`
+   ^^^^^^^^^^^^ role
+    ^^^^^^^^^^ name
+
+The language server sometimes refers to the above as a "plain" role, in that the
+role's target contains just the label of the object it is linking to. However it's
+also possible to define "aliased" roles, where the link text in the final document
+is overriden, for example::
+
+                vvvvvvvvvvvvvvvvvvvvvvvv alias
+                                          vvvvvv label
+                                         v modifier (optional)
+               vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvv target
+   {c:function}`used to allocate memory <~malloc>`
+   ^^^^^^^^^^^^ role
+    ^^^^^^^^^^ name
+
+"""
+
+
+RST_ROLE = re.compile(
+    r"""
+    ([^\w:]|^\s*)                     # roles cannot be preceeded by letter chars
+    (?P<role>
+      :                               # roles begin with a ':' character
+      (?!:)                           # the next character cannot be a ':'
+      ((?P<name>\w([:\w-]*\w)?):?)?   # roles have a name
+    )
+    (?P<target>
+      `                               # targets begin with a '`' character
+      ((?P<alias>[^<`>]*?)<)?         # targets may specify an alias
+      (?P<modifier>[!~])?             # targets may have a modifier
+      (?P<label>[^<`>]*)?             # targets contain a label
+      >?                              # labels end with a '>' when there's an alias
+      `?                              # targets end with a '`' character
+    )?
+    """,
+    re.VERBOSE,
+)
+"""A regular expression to detect and parse parial and complete roles.
+
+I'm not sure if there are offical names for the components of a role, but the
+language server breaks a role down into a number of parts::
+
+                 vvvvvv label
+                v modifier(optional)
+               vvvvvvvv target
+   :c:function:`!malloc`
+   ^^^^^^^^^^^^ role
+    ^^^^^^^^^^ name
+
+The language server sometimes refers to the above as a "plain" role, in that the
+role's target contains just the label of the object it is linking to. However it's
+also possible to define "aliased" roles, where the link text in the final document
+is overriden, for example::
+
+                vvvvvvvvvvvvvvvvvvvvvvvv alias
+                                          vvvvvv label
+                                         v modifier (optional)
+               vvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvvv target
+   :c:function:`used to allocate memory <~malloc>`
+   ^^^^^^^^^^^^ role
+    ^^^^^^^^^^ name
+
+"""
+
+
+RST_DEFAULT_ROLE = re.compile(
+    r"""
+    (?<![:`])
+    (?P<target>
+      `                               # targets begin with a '`' character
+      ((?P<alias>[^<`>]*?)<)?         # targets may specify an alias
+      (?P<modifier>[!~])?             # targets may have a modifier
+      (?P<label>[^<`>]*)?             # targets contain a label
+      >?                              # labels end with a '>' when there's an alias
+      `?                              # targets end with a '`' character
+    )
+    """,
+    re.VERBOSE,
+)
+"""A regular expression to detect and parse parial and complete "default" roles.
+
+A "default" role is the target part of a normal role - but without the ``:name:`` part.
+"""
 
 
 @dataclass
